requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.9
django-simple-history==2.4.0
django-storages==1.7.1
djangorestframework==3.8.2
docutils==0.14
freezegun==0.3.10
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
whitenoise==4.1
